POR-2896: organized folders to be consistent
